add fancy exn handling and tests
import the test-helpers

follow-up on the previous commit message / stream-of-consciousness

this commit contains test cases written even after code coverage was acheived in many spots. (e.g. testing handle-stmt ends up almost fully covering handle-arg becuase handle-arg is very short). Admittedly, testing the fixnum branch again with a number in the fixnum range kind of redundant... but having both `given6` and `given10` tests allows for testing the integration of the funcitons with the fancy `with-handler` syntax, which does some magic to append the current function name to the exception being raised up the call stack.

\end-stream-of-consciousness-again

useful resources

on exceptions:

https://beautifulracket.com/explainer/errors-and-exceptions.html

on string formatting:

https://docs.racket-lang.org/reference/strings.html#%28def._%28%28lib._racket%2Fformat..rkt%29._~7ev%29%29

exn->string

https://docs.racket-lang.org/reference/exns.html#%28def._%28%28lib._racket%2Fexn..rkt%29._exn-~3estring%29%29

on modules:

https://docs.racket-lang.org/guide/module-basics.html