update get-live-after-sets not vectors, seems complete
turns out that vectors are not needed because the L_after variable will always contain the L_after as the first element of the list so accessing that is very easy when building the whole list-of-live-after-set in a bottom-up-recursive-way.

vectors resource...

https://docs.racket-lang.org/reference/vectors.html#(def._((quote._~23~25kernel)._vector))
ALSO sets are cool! fancy functions like set-union and set-subtract make life very easy for calculating live-after-sets

sets resource...

https://docs.racket-lang.org/reference/sets.html